pyproject.toml
[project]
name = "training-treckbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery[redis]>=5.6.2",
    "django>=6.0.2",
    "django-celery-beat>=2.1.0",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "dotenv>=0.9.9",
    "drf-yasg>=1.21.15",
    "gunicorn>=25.1.0",
    "psycopg2-binary>=2.9.11",
    "pytz>=2025.2",
    "rest-framework-simplejwt>=0.0.2",
]

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = ["migrations"]

[tool.ruff.lint]
select = ["E", "W", "F", "I"]
ignore = []

[tool.ruff.format]
quote-style = "double"

[dependency-groups]
dev = [
    "ruff>=0.15.8",
]
